Refactor front-end scripts into modules
