+++ index.html
@@ -204,6 +204,6 @@
       integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
       crossorigin="anonymous"
     ></script>
-    <script src="assets/js/game.js"></script>
+    <script type="module" src="assets/js/game.js"></script>
   </body>
 </html>
